+++ src/datalevin/core.clj
@@ -978,10 +978,11 @@ Only usable for debug output.
 
   `dir` is a directory path or a dtlv connection URI string.
   `opts` is an option map that may have the following keys:
-  * `:mapsize` is the initial size of the database. This will be expanded as needed
+  * `:mapsize` is the initial size of the database. This will be expanded as needed.
   * `:flags` is a vector of keywords corresponding to LMDB environment flags, e.g.
      `:rdonly-env` for MDB_RDONLY_ENV, `:nosubdir` for MDB_NOSUBDIR, and so on. See [LMDB Documentation](http://www.lmdb.tech/doc/group__mdb__env.html)
   * `:temp?` a boolean, indicating if this db is temporary, if so, the file will be deleted on JVM exit.
+  * `:max-dbis` specifies the maximal number of named sub-databases (DBI) that can be opened inside this database. Default is 128. A very large number may induce slowness as DBI lookup is a linear scan.
   * `:client-opts` is the option map passed to the client if `dir` is a remote server URI string.
   * `:spill-opts` is the option map that controls the spill-to-disk behavior for `get-range` and `range-filter` functions, which may have the following keys:
       - `:spill-threshold`, memory pressure in percentage of JVM `-Xmx` (default 80), above which spill-to-disk will be triggered.