+++ test/datalevin/bits_test.clj
@@ -56,19 +56,21 @@
 
 (test/defspec value-compressed-data-generative-test
   100
-  (prop/for-all [k gen/any-equatable]
-                (let [^ByteBuffer bf (bf/allocate-buffer 16384)]
-                  (sut/put-bf bf k :data cp/dict-less-compressor)
-                  (= k (sut/read-buffer bf :data cp/dict-less-compressor)))))
+  (let [compressor (cp/get-dict-less-compressor)]
+    (prop/for-all [k gen/any-equatable]
+                  (let [^ByteBuffer bf (bf/allocate-buffer 16384)]
+                    (sut/put-bf bf k :data compressor)
+                    (= k (sut/read-buffer bf :data compressor))))))
 
 (test/defspec string-generative-test
   100
-  (prop/for-all [k (gen/such-that
-                     (partial string-size-less-than? c/+val-bytes-wo-hdr+)
-                     gen/string)]
-                (let [^ByteBuffer bf (bf/allocate-buffer 16384)]
-                  (sut/put-bf bf k :string cp/dict-less-compressor)
-                  (= k (sut/read-buffer bf :string cp/dict-less-compressor)))))
+  (let [compressor (cp/get-dict-less-compressor)]
+    (prop/for-all [k (gen/such-that
+                       (partial string-size-less-than? c/+val-bytes-wo-hdr+)
+                       gen/string)]
+                  (let [^ByteBuffer bf (bf/allocate-buffer 16384)]
+                    (sut/put-bf bf k :string compressor)
+                    (= k (sut/read-buffer bf :string compressor))))))
 
 (test/defspec int-generative-test
   100
@@ -79,10 +81,11 @@
 
 (test/defspec value-compress-int-generative-test
   100
-  (prop/for-all [k gen/int]
-                (let [^ByteBuffer bf (bf/allocate-buffer 16384)]
-                  (sut/put-bf bf k :int cp/dict-less-compressor)
-                  (= k (sut/read-buffer bf :int cp/dict-less-compressor)))))
+  (let [compressor (cp/get-dict-less-compressor)]
+    (prop/for-all [k gen/int]
+                  (let [^ByteBuffer bf (bf/allocate-buffer 16384)]
+                    (sut/put-bf bf k :int compressor)
+                    (= k (sut/read-buffer bf :int compressor))))))
 
 (test/defspec int-int-generative-test
   100
@@ -325,15 +328,15 @@
 
 (test/defspec compressed-datom-generative-test
   100
-  (prop/for-all [e gen/large-integer
-                 a gen/keyword-ns
-                 v gen/any-equatable
-                 t gen/large-integer]
-                (let [^ByteBuffer bf (bf/allocate-buffer 16384)
-                      d              (d/datom e a v t)]
-                  (sut/put-bf bf d :datom cp/dict-less-compressor)
-                  (is (= d (sut/read-buffer bf :datom
-                                            cp/dict-less-compressor))))))
+  (let [compressor (cp/get-dict-less-compressor)]
+    (prop/for-all [e gen/large-integer
+                   a gen/keyword-ns
+                   v gen/any-equatable
+                   t gen/large-integer]
+                  (let [^ByteBuffer bf (bf/allocate-buffer 16384)
+                        d              (d/datom e a v t)]
+                    (sut/put-bf bf d :datom compressor)
+                    (is (= d (sut/read-buffer bf :datom compressor)))))))
 
 (test/defspec attr-generative-test
   100